5_NLP/requirements.txt
numpy
spark-nlp==2.3.4
google-cloud-storage
google-cloud-bigquery
redis